backend/recruitment-service/requirements.txt
# API and Web Framework
fastapi>=0.103.1
uvicorn>=0.23.2
websockets>=11.0.3
pydantic>=2.3.0
pydantic-settings>=2.0.3

# Database
sqlalchemy>=2.0.20
alembic>=1.12.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.7

# Authentication & Security
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.1

# Event Handling
redis>=5.0.0
aio-pika>=9.2.2

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.3
httpx>=0.24.1
structlog>=23.1.0
aiobotocore>=2.5.4

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
httpx>=0.24.1
asgi-lifespan>=2.1.0

# Documentation
pyyaml>=6.0.1
